requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.13
faiss-cpu==1.7.4
pymavlink==2.4.37
pydantic==2.5.2 